% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{forecast_plots}
\alias{forecast_plots}
\title{Plot forecast summary statistics vs observations}
\usage{
forecast_plots(
  df,
  ncol = NULL,
  show.legend = FALSE,
  width_svg = 8,
  height_svg = 4
)
}
\arguments{
\item{df}{a scored/summarized dataframe from \code{score4cast} or \code{neon4cast::combined_scores}}

\item{ncol}{number of columns on facet_wrap}

\item{show.legend}{show legend?}

\item{width_svg}{svg width}

\item{height_svg}{svg height}
}
\value{
a girafe object
}
\description{
ggiraph interactive javascript plot of forecasts
}
\details{
these functions are designed for the neon4cast dashboard specifically.
They may or may not generalize well to other use cases, consider modifying or
coding directly in ggiraph and ggplot.
}
