% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{leaderboard_plots}
\alias{leaderboard_plots}
\title{Three plots of summary scores: vs model_id overall, vs reference_datetime, and vs horizon}
\usage{
leaderboard_plots(
  df,
  var = NULL,
  horizon_cutoff = 35,
  width_svg = 8,
  height_svg = 6
)
}
\arguments{
\item{df}{a scored/summarized dataframe from \code{score4cast} or \code{neon4cast::combined_scores}}

\item{var}{variable name to filter by (or just pre-filter to one variable yourself)}

\item{horizon_cutoff}{trims forecasts to specified horizon (assumes unit of days)
(Note some teams submit much beyond the horizon requested in the design)}

\item{width_svg}{svg width}

\item{height_svg}{svg height}
}
\value{
girafe object
}
\description{
Three plots of summary scores: vs model_id overall, vs reference_datetime, and vs horizon
}
