% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_scores.R
\name{sync_scores}
\alias{sync_scores}
\title{Sync scores from remote server to a local directory}
\usage{
sync_scores(dest = "cache/")
}
\arguments{
\item{dest}{local directory to cache scores}
}
\description{
Sync scores from remote server to a local directory
}
